+++ src/components/header/DiscoverHeaderButton.js
@@ -3,7 +3,7 @@ import LinearGradient from 'react-native-linear-gradient';
 import styled from 'styled-components';
 import { useNavigation } from '../../navigation/Navigation';
 import { Row, RowWithMargins } from '../layout';
-import { Emoji, Text } from '../text';
+import { Text } from '../text';
 import HeaderButton from './HeaderButton';
 import Routes from '@rainbow-me/routes';
 import { padding, position } from '@rainbow-me/styles';
@@ -13,10 +13,6 @@ const DiscoverButtonShadowsFactory = colors => [
   [0, 7, 21, colors.shadow, 0.06],
   [0, 3.5, 10.5, colors.shadow, 0.04],
 ];
-
-const EmojiWrapper = styled.View`
-  margin-top: ${android ? 6 : 0}px;
-`;
 
 const BackgroundFill = styled.View`
   ${position.cover};
@@ -40,7 +36,7 @@ const DiscoverButtonContent = styled(RowWithMargins).attrs({
 })`
   align-items: center;
   justify-content: center;
-  ${padding(ios ? 6 : 0, 10, 7.5)};
+  ${padding(2, 10, 7.5)};
   height: 34;
   z-index: 2;
 `;
@@ -83,16 +79,14 @@ export default function DiscoverHeaderButton() {
           <BackgroundGradient />
         </ShadowStack>
         <DiscoverButtonContent>
-          <EmojiWrapper>
-            <Emoji size="small">🪐</Emoji>
-          </EmojiWrapper>
           <Text
             color={colors.alpha(colors.blueGreyDark, isDarkMode ? 1 : 0.8)}
             letterSpacing="roundedTight"
             size="large"
             weight="bold"
+            {...(android && { lineHeight: 32 })}
           >
-            Discover
+            🪐 Discover
           </Text>
         </DiscoverButtonContent>
       </Row>