+++ src/components/list/ListHeader.js
@@ -19,7 +19,7 @@ import {
 import { RAINBOW_PROFILES_BASE_URL } from '@rainbow-me/references';
 import { padding, position } from '@rainbow-me/styles';
 
-export const ListHeaderHeight = 44;
+export const ListHeaderHeight = 50;
 
 const BackgroundGradient = styled(LinearGradient).attrs(
   ({ theme: { colors } }) => ({
@@ -37,21 +37,20 @@ const BackgroundGradient = styled(LinearGradient).attrs(
 `;
 
 const ShareCollectiblesBPA = styled(ButtonPressAnimation)`
-  background-color: ${({ theme: { colors } }) => colors.blueGreyDarkLight};
-  margin-right: 0;
-  width: 90;
-  height: 30;
+  background-color: ${({ theme: { colors } }) =>
+    colors.alpha(colors.blueGreyDark, 0.06)};
   border-radius: 15;
-  padding-left: 12;
-  padding-right: 8;
-  padding-top: 5;
-  padding-bottom: 5;
+  height: 30;
   justify-content: center;
+  max-width: 90;
+  padding-bottom: 5;
+  padding-top: 5;
+  width: 90;
 `;
 
 const ShareCollectiblesButton = ({ onPress }) => (
   <ShareCollectiblesBPA onPress={onPress} scale={0.9}>
-    <CoinDividerButtonLabel label="􀈂 Share" />
+    <CoinDividerButtonLabel align="center" label="􀈂 Share" shareButton />
   </ShareCollectiblesBPA>
 );
 
@@ -59,7 +58,7 @@ const Content = styled(Row).attrs({
   align: 'center',
   justify: 'space-between',
 })`
-  ${padding(0, 19, 2)};
+  ${padding(5, 19)};
   background-color: ${({ isSticky, theme: { colors } }) =>
     isSticky ? colors.white : colors.transparent};
   height: ${ListHeaderHeight};
@@ -84,6 +83,7 @@ export default function ListHeader({
   totalValue,
 }) {
   const deviceDimensions = useDimensions();
+  const { colors } = useTheme();
   const { isReadOnlyWallet } = useWallets();
   const { accountAddress } = useAccountSettings();
   const { accountENS } = useAccountProfile();
@@ -125,20 +125,24 @@ export default function ListHeader({
       <Fragment>
         <BackgroundGradient />
         <Content isSticky={isSticky}>
-          <Row align="center">
-            {createElement(titleRenderer, { children: title })}
-            {title === 'Collectibles' && (
-              <Column align="flex-end" flex={1}>
-                <ShareCollectiblesButton
-                  onPress={() => handleShare(isReadOnlyWallet, accountAddress)}
-                />
-              </Column>
-            )}
-            <ContextMenu marginTop={3} {...contextMenuOptions} />
-          </Row>
+          {title && (
+            <Row align="center">
+              {createElement(titleRenderer, { children: title })}
+              {title === 'Collectibles' && (
+                <Column align="flex-end" flex={1}>
+                  <ShareCollectiblesButton
+                    onPress={() =>
+                      handleShare(isReadOnlyWallet, accountAddress)
+                    }
+                  />
+                </Column>
+              )}
+              <ContextMenu marginTop={3} {...contextMenuOptions} />
+            </Row>
+          )}
           {children}
         </Content>
-        {showDivider && <Divider />}
+        {showDivider && <Divider color={colors.rowDividerLight} />}
         {!isSticky && title !== 'Balances' && (
           <StickyBackgroundBlocker
             deviceDimensions={deviceDimensions}